backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
vaderSentiment==3.3.2
google-generativeai==0.7.2
elevenlabs==1.5.0
python-dotenv==1.0.1
PyJWT==2.8.0
bcrypt==4.1.3
psycopg2-binary==2.9.9
